[batch] add progress helper and use in sow and till
